backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "generating TC"
authors = ["Farah Ouesleti <farah.oueslati@talan.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pandas = "^2.2.3"
load-dotenv = "^0.1.0"
jira = "^3.8.0"
openai = "^1.55.1"
autogen = "^0.4"
graphrag = "^0.5.0"
fitz = "^0.0.1.dev2"
langchain-core = "^0.3.21"
pymupdf = "^1.24.14"
python-multipart = "^0.0.17"
openpyxl = "^3.1.5"
fastapi-cors = "^0.0.6"
file = "^0.3.0"
form = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
